usergroup resource fixed update id "(known after apply)"
This was causing any resources using this id to be recreated, which is
not required as the ID doesn't change.

Added "resource.UseStateForUnknown() attribute plan modifier" as
recommended here https://developer.hashicorp.com/terraform/plugin/framework/resources/update#recommendations
to optional attributes that do not need to re-computed on update if they
are not set.